Coding 2025-2026/Security/7/sql-injection-lab/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-dotenv
bcrypt
pydantic
pytest
httpx
pytest-asyncio
aiosqlite
asgi-lifespan
SQLAlchemy
alembic